// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AccountType = {
  NEQUI: 'NEQUI',
  DAVIPLATA: 'DAVIPLATA',
  BANCOLOMBIA: 'BANCOLOMBIA',
  NUBANK: 'NUBANK',
  EFECTIVO: 'EFECTIVO',
  OTRO: 'OTRO'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]


export const TransactionType = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE',
  TRANSFER: 'TRANSFER'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const CategoryType = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE'
} as const

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType]